requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
